Erstes Refactoring durch
