+++ Gamefiles/game.py
@@ -2,11 +2,14 @@ import pygame
 import math
 import sys
 
-from Util import Util
-from Player import Player
-from Background import Background
-from Road import Road
+from Gamefiles.util import Util
+from Visuals.player import Player
+from Visuals.background import Background
+from Visuals.road import Road
+from Visuals.colors import Colors
 
+# Hiermit bildet man die Strasse. Es wird wahrscheinlich wichtig sein, dass dieses Array genauso auch beim Java-Client aufgebaut ist.
+# TODO: Gemeinsame Speicherloesung finden
 road = [
     [100, 0, 60],
     [50],
@@ -16,55 +19,24 @@ road = [
     [90, -4]
 ]
 
-def get_finish():
-        c = {
-            "road": "#000000",
-            "grass": "#000000",
-            "rumble": "#000000",
-            "lane": "#000000"
-        }
-        return c
-
-def get_light():
-        c = {
-            "road": "#6B6B6B",
-            "grass": "#10AA10",
-            "rumble": "#555555",
-            "lane": "#CCCCCC"
-        }
-        return c
-
-def get_start():
-        c = {
-            "road": "#FFFFFF",
-            "grass": "#FFFFFF",
-            "rumble": "#FFFFFF",
-            "lane": "#FFFFFF"
-        }
-        return c
-
-def get_dark():
-        c = {
-            "road": "#696969",
-            "grass": "#009A00",
-            "rumble": "#BBBBBB",
-            "lane": "#696969"
-        }
-        return c
 # Der Teil des Singleplayers, der alle anderen aktiviert: Hier werden Tasteneingaben überprüft und die Straße unendlich erweitert
 class Game:
 
     # Sowohl in Python als auch in JavaScript sorgt das Veraedern der FPS fuer Probleme beim Spielverhalten
     FPS = 60
-    step = 1/FPS
-    dt = 1/FPS/2
+    STEP = 1/FPS
+    DT = 1/FPS/2
+
     width = 1024
     height = 768
+
     segments = []
+
     screen = None
     background = None
     sprites = None
     resolution = None
+
     # roadWidth bestimmt, wie breit die gesamte Strasse (alle Spuren) sein soll
     roadWidth = 2000
     segmentLength = 200
@@ -72,22 +44,27 @@ class Game:
     trackLength = 0
     # lanes bestimmt, wie viele Spuren es auf der Straße geben soll (je breiter die Straße, desto mehr Spuren können realistisch genutzt werden)
     lanes = 3
+
     fov = 100
     cameraHeight = 1000
     cameraDepth = 1 / math.tan((fov / 2) * math.pi / 180)
     drawDistance = 300
     playerX = 0
     playerZ = cameraHeight * cameraDepth
-    fogDensity = 5
+    fogDensity = 10
+
     position = 0
     speed = 0
-    maxSpeed = segmentLength/step
-    accel = maxSpeed/5
+    maxSpeed = segmentLength/STEP
+    accel = maxSpeed/5 - 10
     breaking = -maxSpeed
     decel = -maxSpeed/5
     offRoadDecel = -maxSpeed/2
     offRoadLimit = maxSpeed/4
+
+    # Beeinflusst das Verhalten des Autos in der Kurve; Je groesser der Wert, desto schwieriger die Lenkung
     centrifugal = 0.3
+    # Falls wir V4 vom JavaScript machen wollen, werden die NPC-Autos in diesem Array gespeichert
     cars = []
 
     keyLeft = False
@@ -99,6 +76,7 @@ class Game:
     def __init__(self):
         pygame.init()
         self.screen = pygame.display.set_mode((self.width, self.height))
+        pygame.display.set_caption("Singleplayer")
         self.player_sprites = pygame.sprite.Group()
         self.background_sprites = pygame.sprite.Group()
         self.reset_road()
@@ -110,9 +88,7 @@ class Game:
         self.create_background()
         self.background_sprites.draw(self.screen)
 
-
         while True:
-            
             for event in pygame.event.get():
                 if event.type == pygame.QUIT:
                     pygame.quit()
@@ -137,7 +113,7 @@ class Game:
                         self.keySlower = False
 
             self.render()
-            self.update(self.step)
+            self.update(self.STEP)
             pygame.display.update()
 
     # Hier wird anhand der Nutzereingaben die Steuerung des Autos geaendert
@@ -161,35 +137,36 @@ class Game:
         self.playerX -= tilt * (self.speed / self.maxSpeed) * current_segment.get("curve") * self.centrifugal
 
         if self.keyFaster:
-            self.speed = Util.accelerate(self.speed, self.accel, self.dt)
+            self.speed = Util.accelerate(self.speed, self.accel, self.DT)
         elif self.keySlower:
-            self.speed = Util.accelerate(self.speed, self.breaking, self.dt)
+            self.speed = Util.accelerate(self.speed, self.breaking, self.DT)
         else:
-            self.speed = Util.accelerate(self.speed, self.decel, self.dt)
+            self.speed = Util.accelerate(self.speed, self.decel, self.DT)
 
         if (self.playerX < -1 or self.playerX > 1) and (self.speed > self.offRoadLimit):
-            self.speed = Util.accelerate(self.speed, self.offRoadDecel, self.dt)
+            self.speed = Util.accelerate(self.speed, self.offRoadDecel, self.DT)
 
         self.playerX = Util.limit(self.playerX, -2, 2)
         self.speed = Util.limit(self.speed, 0, self.maxSpeed)
 
         playerW = ((1/80) * 0.3) * 80
         
-
+    # Liest das Strassen-Array aus und markiert Start-/Ziellinie
     def reset_road(self):
-        self.segments = []
 
         self.read_road()
 
-        self.segments[self.which_segment(self.playerZ)["index"] + 2]["color"] = get_start()
-        self.segments[self.which_segment(self.playerZ)["index"] + 3]["color"] = get_start()
+        self.segments[self.which_segment(self.playerZ)["index"] + 2]["color"] = Colors.get_start()
+        self.segments[self.which_segment(self.playerZ)["index"] + 3]["color"] = Colors.get_start()
         for n in range(self.rumbleLength):
-            self.segments[len(self.segments)-1-n]["color"] = get_finish()
+            self.segments[len(self.segments)-1-n]["color"] = Colors.get_finish()
 
         self.trackLength = len(self.segments) * self.segmentLength
-        
+    
+    # Hier wird das hinzuzufuegende Segment anhand von Kurven- und Huegel-Parametern angepasst
     def add_segment(self, curve, y):
         n = len(self.segments)
+        # Diese Werte muessen (hoffentlich) nicht mehr geaendert werden, jetzt wo Kurven (curve-Parameter) und Huegel (y-Parameter) benutzt werden
         self.segments.append(
                 {
                     'index': n,
@@ -234,6 +211,7 @@ class Game:
                 }
             )
 
+    # Fuegt Strassenteile an das Segment-Array hinzu
     def add_road(self, enter, hold, leave, curve, y=0):
         startY = self.lastY()
         endY = startY + (int(y)*self.segmentLength)
@@ -248,6 +226,7 @@ class Game:
         for n in range(int(leave)):
             self.add_segment(Util.easeInOut(0, curve, n / enter), Util.easeInOut(startY, endY, (enter + hold + n) / total))
 
+    # Ueberprueft, ob Werte des Straßenmoduls leer ist
     def add_street(self, num=None, curve=None, height=None):
         if num is None:
             num = Road.length().get("medium")
@@ -258,6 +237,7 @@ class Game:
 
         self.add_road(num, num, num, curve, height)
 
+    # Gibt die letzte Y-Koordinate aus, um einen glatten Huegel zu modellieren
     def lastY(self):
         if len(self.segments) == 0:
             return 0
@@ -266,9 +246,9 @@ class Game:
     # Hilfsmethode, um zu gucken, welche Farbe das aktuelle Strassenstueck haben muss
     def which_road(self, n):
          if (n / self.rumbleLength) % 2 == 0:
-              return get_light()
+              return Colors.get_light()
          else:
-              return get_dark()
+              return Colors.get_dark()
 
     # Hilfsmethode, um das derzeit notwendige Segment auszuwaehlen
     def which_segment(self, n):
@@ -276,19 +256,19 @@ class Game:
     
     # Setzt ab einem bestimmten Punkt die Distanz der Strecke zurueck (?) und zeichnet Strecke, Nebel und Spieler auf den Bildschirm
     def render(self):
-         base = self.which_segment(self.position)
-         base_percent = Util.percent_remaining(self.position, self.segmentLength)
-         current_segment = self.which_segment(self.position + self.playerZ)
-         current_percent = Util.percent_remaining(self.position + self.playerZ, self.segmentLength)
-         playerY = Util.interpolate(current_segment.get("p1").get("world").get("y"), current_segment.get("p2").get("world").get("y"), current_percent)
+        base = self.which_segment(self.position)
+        base_percent = Util.percent_remaining(self.position, self.segmentLength)
+        current_segment = self.which_segment(self.position + self.playerZ)
+        current_percent = Util.percent_remaining(self.position + self.playerZ, self.segmentLength)
+        playerY = Util.interpolate(current_segment.get("p1").get("world").get("y"), current_segment.get("p2").get("world").get("y"), current_percent)
 
-         dx = -(base.get("curve") * base_percent)
-         x = 0
-         maxY = self.height
+        dx = -(base.get("curve") * base_percent)
+        x = 0
+        maxY = self.height
 
-         self.background_sprites.draw(self.screen)
+        self.background_sprites.draw(self.screen)
 
-         for n in range(self.drawDistance):
+        for n in range(self.drawDistance):
             segment = self.segments[(base.get("index") + n) % len(self.segments)]
             segment_looped = segment.get("index") < base.get("index")
             segment_fog = Util.exponential_fog(n / self.drawDistance, self.fogDensity)
@@ -326,13 +306,13 @@ class Game:
                 continue
 
             Util.segment(self.screen, self.width, self.lanes,
-                         segment.get("p1").get("screen").get("x"),
-                         segment.get("p1").get("screen").get("y"),
-                         segment.get("p1").get("screen").get("w"),
-                         segment.get("p2").get("screen").get("x"),
-                         segment.get("p2").get("screen").get("y"),
-                         segment.get("p2").get("screen").get("w"),
-                         segment.get("color"), segment_fog)
+                        segment.get("p1").get("screen").get("x"),
+                        segment.get("p1").get("screen").get("y"),
+                        segment.get("p1").get("screen").get("w"),
+                        segment.get("p2").get("screen").get("x"),
+                        segment.get("p2").get("screen").get("y"),
+                        segment.get("p2").get("screen").get("w"),
+                        segment.get("color"), segment_fog)
 
             maxY = segment.get("p1").get("screen").get("y")
 
@@ -353,6 +333,7 @@ class Game:
         self.player = Player(self.screen.get_width() / 2 - 30, self.screen.get_height() - 100)
         self.player_sprites.add(self.player)
 
+    # Liest das Strassen-Array aus und entscheidet je nach Menge an Werten, ob Kurven oder Huegel zum Segment gehoeren
     def read_road(self):
         for x in road:
             if len(x) == 1:
